Large Processing Update
Significant steps have been made in processing the data. There are two methods for alligning the lidar data, either using the time the lidar was acquired and spreading it across the LSP scan at that time evenly, or using the angle of the lidar and placing the distance value in the appropriate LSP position. The latter is advised. post_process.py runs and works, plotting the data for  singel angle for both the LSP and lidar. Tests on an arm moving show that it works, with a lag associated with instrument offset - correction for this has not yet been developed