GUI started
Started to build a GUI to make processing more user friendly. Works for processing current data, and has the option to change instrument orientation. Need to matplotlib navigation toolbars to figures and then work on introducing more settings etc. Also then need to build the acquisition side of the gui